+++ docs/src/examples/components/Popup/Types/PopupExample.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Button, Popup } from '@stardust-ui/react'
+
+const PopupExample = () => <Popup trigger={<Button icon="expand" />}>Add users to your feed</Popup>
+
+export default PopupExample